Move optimize files to tinyopt
