+++ include/tinyopt/optimizers/gd.h
@@ -15,7 +15,7 @@
 #pragma once
 
 #include <tinyopt/math.h>
-#include <tinyopt/optimizers/optimize.h>
+#include <tinyopt/optimize.h>
 
 #include <tinyopt/solvers/gd.h>
